Settings button and panel per scene
Settings button now works in each scene. Audio still does not play - possible deletion of wwise as this conflicts with one another.
