+++ Assets/Prefabs/Settings Panel.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 9b4758efba4292543af7bc2b1b45aa51
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
