+++ Assets/Scripts/Managers/BasePanelManager.cs
@@ -1,9 +1,12 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using UnityEngine.UI;
 
 public class BasePanelManager : MonoBehaviour
 {
+    public GameObject settingsPanel;
+    public Slider volumeSlider;
     public void LoadScene(string sceneName)
     {
         if (SceneManagement.Instance != null)
@@ -14,5 +17,15 @@ public class BasePanelManager : MonoBehaviour
         {
             Debug.LogError("SceneManagement instance is not available.");
         }
+    }
+
+    public void OpenSettings()
+    {
+        settingsPanel.SetActive(true);
+    }
+
+    public void CloseSettings()
+    {
+        settingsPanel.SetActive(false);
     }
 }
